Extract retirement dashboard components and add layout-aware skeleton
- Splits the retirement dashboard page into focused components:
  retirement-portfolio-chart, retirement-coverage-chart,
  retirement-snapshot-table, and value-mode-toggle
- Adds retirement-dashboard-skeleton that mirrors the real two-column
  layout (hero, chart, milestone strip, coverage, snapshot, sidebar)
  instead of three generic rectangles
- Adopts the shared formatCompactAmount helper across dashboard and
  risk-lab pages
- Updates guide-page pension copy to mention country-agnostic examples
  (CPP/OAS in Canada, Social Security in the US)
